Fix Bugs from Test Result
Bugs:
- Wrong source of decorators in minierp/views (was referencing hardware)
- load_building in employee views instead of hardware views
- Order of url patterns checking for employee detail page pattern when
running test for employee add (first checks 'employee/x' pattern > found
x should be an pk, but request was x to be add/, changed the order to
fix this)
- employee/test_views EmployeeDashboard test checking base url ('employee')
instead of dashboard ('/')
