+++ employee/tests/test_views.py
@@ -32,7 +32,7 @@ class HomePageTest(LoginTestCase):
 class EmployeeDashboardPageTests(LoginTestCase):
 
     def test_url_exists_at_correct_location(self):
-        response = self.client.get("/employees/")
+        response = self.client.get("/")
         self.assertEqual(response.status_code, 200)
     
     def test_url_available_by_name(self):
